+++ patches/src/main/kotlin/dev/jkcarino/adobo/patches/google/gboard/detection/signature/BypassSignaturePatch.kt
@@ -2,12 +2,13 @@ package dev.jkcarino.adobo.patches.google.gboard.detection.signature
 
 import app.morphe.patcher.patch.bytecodePatch
 import dev.jkcarino.adobo.patches.google.gboard.fixes.applyWorkaroundPatch
+import dev.jkcarino.adobo.patches.google.gboard.shared.COMPATIBILITY_GBOARD
 import dev.jkcarino.adobo.util.returnEarly
 
 val bypassSignaturePatch = bytecodePatch(
     description = "Bypasses the signature verification checks."
 ) {
-    compatibleWith("com.google.android.inputmethod.latin")
+    compatibleWith(COMPATIBILITY_GBOARD)
 
     dependsOn(applyWorkaroundPatch)