+++ patches/src/main/kotlin/dev/jkcarino/adobo/patches/google/gboard/incognito/clipboard/EnableClipboardPatch.kt
@@ -3,14 +3,14 @@ package dev.jkcarino.adobo.patches.google.gboard.incognito.clipboard
 import app.morphe.patcher.extensions.InstructionExtensions.removeInstructions
 import app.morphe.patcher.patch.bytecodePatch
 import dev.jkcarino.adobo.patches.google.gboard.detection.signature.bypassSignaturePatch
+import dev.jkcarino.adobo.patches.google.gboard.shared.COMPATIBILITY_GBOARD
 
 @Suppress("unused")
 val enableClipboardPatch = bytecodePatch(
     name = "Enable clipboard in incognito",
-    description = "Enables clipboard support in incognito mode.",
-    use = true
+    description = "Enables clipboard support in incognito mode."
 ) {
-    compatibleWith("com.google.android.inputmethod.latin")
+    compatibleWith(COMPATIBILITY_GBOARD)
 
     dependsOn(bypassSignaturePatch)