+++ patches/src/main/kotlin/dev/jkcarino/adobo/patches/reddit/layout/actions/share/Fingerprints.kt
@@ -1,6 +1,7 @@
 package dev.jkcarino.adobo.patches.reddit.layout.actions.share
 
 import app.morphe.patcher.Fingerprint
+import dev.jkcarino.adobo.patches.reddit.shared.LinkToStringFingerprint
 
 internal object ActionCellFragmentToStringFingerprint : Fingerprint(
     returnType = "Ljava/lang/String;",
@@ -12,6 +13,7 @@ internal object ActionCellFragmentToStringFingerprint : Fingerprint(
 )
 
 internal object GetShareCountFingerprint : Fingerprint(
+    classFingerprint = LinkToStringFingerprint,
     name = "getShareCount",
     returnType = "Ljava/lang/Long;",
     parameters = listOf()