+++ patches/src/main/kotlin/dev/jkcarino/adobo/patches/reddit/misc/outboundlink/Fingerprints.kt
@@ -1,30 +1,34 @@
 package dev.jkcarino.adobo.patches.reddit.misc.outboundlink
 
 import app.morphe.patcher.Fingerprint
+import dev.jkcarino.adobo.patches.reddit.shared.LinkToStringFingerprint
 
-internal object AccountPreferencesToStringFingerprint : Fingerprint(
+private object AccountPreferencesToStringFingerprint : Fingerprint(
     returnType = "Ljava/lang/String;",
     parameters = listOf(),
     strings = listOf("AccountPreferences(over18=")
 )
 
-internal object GetAllowClickTrackingFingerprint : Fingerprint(
-    name = "getAllowClickTracking",
-    returnType = "Z"
-)
-
-internal object AccountToStringFingerprint : Fingerprint(
+private object AccountToStringFingerprint : Fingerprint(
     returnType = "Ljava/lang/String;",
     parameters = listOf(),
     strings = listOf("Account(id=")
 )
 
+internal object GetAllowClickTrackingFingerprint : Fingerprint(
+    classFingerprint = AccountPreferencesToStringFingerprint,
+    name = "getAllowClickTracking",
+    returnType = "Z"
+)
+
 internal object GetOutboundClickTrackingFingerprint : Fingerprint(
+    classFingerprint = AccountToStringFingerprint,
     name = "getOutboundClickTracking",
     returnType = "Z"
 )
 
 internal object GetOutboundLinkFingerprint : Fingerprint(
+    classFingerprint = LinkToStringFingerprint,
     name = "getOutboundLink",
     returnType = "L",
     parameters = listOf()